requirements.txt
streamlit
pandas
matplotlib
seaborn
missingno
scikit-learn
numpy
flask
flask-mysql
mysql-connector-python

